feat: implemented navigation between screens in MainActivity
