+++ app/src/main/java/com/example/valorant_app/ui/pages/agent/AgentsScreen.kt
@@ -1,0 +1,13 @@
+package com.example.valorant_app.ui.pages.agent
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.navigation.NavController
+
+@Composable
+fun AgentsScreen(
+    navController: NavController,
+    modifier: Modifier = Modifier
+) {
+
+}
