+++ app/src/main/java/com/example/valorant_app/ui/pages/home/HomePageScreen.kt
@@ -1,0 +1,12 @@
+package com.example.valorant_app.ui.pages
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavController
+import com.example.valorant_app.ui.pages.home.HomeContent
+
+@Composable
+fun HomePageScreen(
+    navController: NavController
+) {
+    HomeContent()
+}
