+++ app/src/main/java/com/example/valorant_app/ui/pages/skin/SkinsScreen.kt
@@ -1,0 +1,13 @@
+package com.example.valorant_app.ui.pages.skin
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.navigation.NavController
+
+@Composable
+fun SkinsScreen(
+    navController: NavController,
+    modifier: Modifier = Modifier
+) {
+
+}
