+++ app/src/main/java/com/example/valorant_app/ui/reusable_comp/SectionCard.kt
@@ -1,0 +1,55 @@
+package com.example.valorant_app.ui.reusable_comp
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.*
+import androidx.compose.material3.DividerDefaults
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun SectionCard(
+    title: String,
+    content: String? = null,
+    accent: Color,
+    contentBlock: (@Composable ColumnScope.() -> Unit)? = null
+) {
+    Card(
+        shape = RoundedCornerShape(8.dp),
+        colors = CardDefaults.cardColors(containerColor = Color(0xFF191A1C)),
+        modifier = Modifier.fillMaxWidth()
+    ) {
+        Column(modifier = Modifier.padding(16.dp)) {
+            Row(verticalAlignment = Alignment.CenterVertically) {
+                HorizontalDivider(
+                    modifier = Modifier
+                        .width(4.dp)
+                        .height(24.dp),
+                    thickness = DividerDefaults.Thickness,
+                    color = accent
+                )
+                Spacer(modifier = Modifier.width(8.dp))
+                Text(title, style = MaterialTheme.typography.titleMedium, color = Color.White)
+            }
+            Spacer(modifier = Modifier.height(8.dp))
+            content?.let {
+                Text(it, style = MaterialTheme.typography.bodyMedium, color = Color.LightGray)
+            }
+            contentBlock?.invoke(this)
+        }
+    }
+}
+
+@Composable
+fun FeatureRow(icon: ImageVector, text: String) {
+    Row(verticalAlignment = Alignment.CenterVertically, modifier = Modifier.padding(vertical = 4.dp)) {
+        Icon(icon, contentDescription = text, modifier = Modifier.size(20.dp), tint = Color(0xFFFF4655))
+        Spacer(modifier = Modifier.width(8.dp))
+        Text(text, style = MaterialTheme.typography.bodySmall, color = Color.White)
+    }
+}
